pyproject.toml
[project]
name = "webwatchdog"
version = "1.0.0"
description = "Webpage checker"
requires-python = ">=3.11"
dependencies = [
    "apscheduler>=3.11.0",
    "email-validator>=2.2.0",
    "flask>=3.1.0",
    "flask-sqlalchemy>=3.1.1",
    "psycopg2-binary>=2.9.10",
    "python-dotenv>=1.0.1",
    "python-telegram-bot>=21.9",
    "requests>=2.32.3",
    "sqlalchemy>=2.0.36",
    "trafilatura>=2.0.0",
]
